requirements.txt
pycryptodome
python-dotenv
PySocks